+++ src/main/java/com/projeto/demo/repositories/projections/TrackCountProjection.java
@@ -1,7 +1,7 @@
 package com.projeto.demo.repositories.projections;
 
 public interface TrackCountProjection {
-    Integer getTrackId();
+    Long getTrackId();
     String getTrackName();
     Long getTotal();
 }
